Replace NewLexer with Lexer
