+++ lib/lrama.rb
@@ -16,4 +16,3 @@ require "lrama/states_reporter"
 require "lrama/type"
 require "lrama/version"
 require "lrama/warning"
-require "lrama/new_lexer"
